refactor: rename provider component
