+++ src/pages/CarDetails/CarDetails.jsx
@@ -4,7 +4,7 @@ import Helmet from "../../components/Helmet/Helmet";
 import { Container, Row, Col } from "reactstrap";
 import { useParams } from "react-router-dom";
 import CarForSale from "../../assets/images/car-for-sale.png";
-import { GetListContext } from "../../context/getCarList";
+import { GetListContext } from "../../context/ListProvider";
 
 const CarDetails = () => {
   const { cars } = useContext(GetListContext);
